pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plantdoc"
version = "0.1.0"
description = "Deep Learning project for multiclass CNN with CBAM"
readme = "README.md"
requires-python = ">=3.8,<3.12"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Typing :: Typed",
]
dependencies = [
  "typer>=0.9.0",
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "optuna>=3.4.0",
  "matplotlib>=3.7.2",
  "seaborn>=0.13.0",
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "jinja2>=3.1.2",
  "scikit-learn>=1.3.0",
  "albumentations>=1.3.1",
  "numpy>=1.24.3",
  "pandas>=2.0.3",
  "pillow>=10.0.0",
  "tqdm>=4.66.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov",
  "black>=23.0.0",
  "isort>=5.10.0",
  "ruff>=0.1.0",
  "mypy>=1.0.0",
  "pre-commit",
]

[project.urls]
Repository = "https://github.com/yourusername/plantdoc"

# --- Tool Configurations ---
[tool.setuptools]
packages = ["plantdoc"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py38"

select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "SIM", # flake8-simplify
]
ignore = [
  "E501", # Line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.8"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
ignore_missing_imports = true
check_untyped_defs = true
pretty = true
